JATZ MAIL VERSION 0.2
We have added special symbols, a dev section with a password that is "hidden" in the code. The 2020 simulator takes you to an interesting youtube video, and a rating feature as well. There is a rare chance of getting a special version of JATZ mail if you open it enough times. All in all we hope you enjoy!